Move commonjs to cli
Update rollup commonjs package to 25
