+++ packages/cli/package.json
@@ -34,6 +34,7 @@
     "cli:debug": "node -r ts-node/register/transpile-only -r tsconfig-paths/register --inspect-brk src/index.ts"
   },
   "dependencies": {
+    "@rollup/plugin-commonjs": "^25.0.0",
     "@stoplight/json": "~3.21.0",
     "@stoplight/path": "1.3.2",
     "@stoplight/spectral-core": "^1.18.3",
