+++ packages/ruleset-bundler/package.json
@@ -35,6 +35,7 @@
     "url": "https://github.com/stoplightio/spectral.git"
   },
   "dependencies": {
+    "@rollup/plugin-commonjs": "^25.0.0",
     "@stoplight/path": "1.3.2",
     "@stoplight/spectral-core": ">=1",
     "@stoplight/spectral-formats": ">=1",
@@ -52,7 +53,7 @@
     "validate-npm-package-name": "3.0.0"
   },
   "devDependencies": {
-    "@rollup/plugin-commonjs": "~22.0.2",
+    "@stoplight/spectral-cli": "workspace:^",
     "@types/validate-npm-package-name": "^3.0.3",
     "fetch-mock": "^9.11.0",
     "memfs": "^3.3.0",
